#378 Исправил ошибки обновления
